Replace every git_gerrit occurence to gerrit_tui
