+++ setup.cfg
@@ -21,4 +21,4 @@ where = src
 
 [options.entry_points]
 console_scripts =
-    gerrit-tui = git_gerrit.__main__:main
+    gerrit-tui = gerrit_tui.__main__:main
