+++ src/gerrit_tui/action/CommentAction.py
@@ -4,8 +4,8 @@ import os
 import json
 import urwid
 
-from git_gerrit.model.Button import Button
-from git_gerrit.model.InputHandler import InputHandler
+from gerrit_tui.model.Button import Button
+from gerrit_tui.model.InputHandler import InputHandler
 
 
 class CommentAction(urwid.WidgetWrap):
