+++ src/gerrit_tui/action/GitResetPrevAction.py
@@ -1,19 +1,19 @@
 import urwid
 
-from git_gerrit.model.Git import Git
-from git_gerrit.model.Button import Button
-from git_gerrit.model.ErrorPopup import display
+from gerrit_tui.model.Git import Git
+from gerrit_tui.model.Button import Button
+from gerrit_tui.model.ErrorPopup import display
 
 
-class GitResetHeadAction(urwid.WidgetWrap):
+class GitResetPrevAction(urwid.WidgetWrap):
     def __init__(self, chageview):
         self.cview = chageview
-        super().__init__(Button("Reset HEAD", "button", self.reset))
+        super().__init__(Button("Reset HEAD~1", "button", self.reset))
 
     def reset(self, w):
         git = Git()
         try:
             git.fetch(self.cview.change['revisions'][self.cview.active_revision_sha]['ref'])
-            git.reset("HEAD")
+            git.reset("HEAD~1")
         except Exception as e:
             display(self.cview.main, str(e))
