+++ src/gerrit_tui/model/ChangeList.py
@@ -1,7 +1,7 @@
 import urwid
 
-from git_gerrit.model.SelectableListItem import SelectableListItem
-from git_gerrit.model.Timestamp import Timestamp
+from gerrit_tui.model.SelectableListItem import SelectableListItem
+from gerrit_tui.model.Timestamp import Timestamp
 
 
 class ChangeList(urwid.ListBox):
