+++ src/gerrit_tui/model/Plugin.py
