+++ src/gerrit_tui/model/SelectableListItem.py
